homework

+++ cart/internal/models/cart.go
@@ -4,6 +4,6 @@ type CartItem struct {
 	UserID int64
 	SKU    uint32
 	Count  int16
-	Price  uint32
+	Price  float64
 	Stock  int16
 }
