+++ stocks/internal/delivery/dto.go
@@ -3,11 +3,11 @@ package delivery
 import "stocks/internal/models"
 
 type StockItemDTO struct {
-	UserID   int64  `json:"user_id"`
-	SKU      uint32 `json:"sku"`
-	Price    uint32 `json:"price"`
-	Count    uint16 `json:"count"`
-	Location string `json:"location"`
+	UserID   int64   `json:"user_id"`
+	SKU      uint32  `json:"sku"`
+	Price    float64 `json:"price"`
+	Count    uint16  `json:"count"`
+	Location string  `json:"location"`
 }
 
 func (dto *StockItemDTO) ToModel() models.StockItem {
