+++ stocks/internal/repository/stock_item_row.go
@@ -6,7 +6,7 @@ type StockItemRow struct {
 	SKU      uint32
 	Name     string
 	Type     string
-	Price    uint32
+	Price    float64
 	Count    uint16
 	Location string
 }
